api/firmware: reduce lowest supported version for MEW
The breaking change from 8.0.0 to 9.0.0 affects only BTC. We reduce
the lowest number to be able to release a new version of the NPM
package using bitbox02-api-js, for use at MyEtherWallet, which will
work fine with 8.0.0 and avoids any downtime for users. Otherwise we
would have to time the firmware release with the MEW release exactly,
which is hard.