api/firmware/query: log time since last query in case of error
The BitBox02 (currently at v9.14.0) aborts a workflow like the
password entry after 500ms of no communciation from the host. In our
polling loop, we query every 200ms.

We have received multiple user reports that on Windows, the BitBoxApp
shows the password entry screen (meaning the device was recognized),
but the device shows the "See the BitBoxApp" screen. At least one user
reported a "unexpected NACK response". Maybe for some reason, the
polling loop is interrupted for too long.

We log the time since the last query to see if that might be the
reason, and how long such an interrupt could be.