backup: don't fire Initialized event when creating backup again
For the initial setup, we need to change the status to Initialized
after the backup was created. When creating the backup again
afterwards, there is no need. The BitBoxApp for example uses this to
register the keystore and load accounts, which does not need to happen
when the accounts are already loaded and the user simply creates a new
backup.