api/firmware: update protobuf msg to support Bitcoin wallet policies
This allows using the policy API (i.e. descriptors containing miniscript).

A new example cli utility was added which registers a policy,
generates a receive address and signs a transcation.

The miniscript-go library and btcd are used in the example tool to
verify the receive address is correct (matches the independently
derived address) and that the generated transaction witness and
sigantures are correct.