firmware/query: improve last usb query intervals log
The previous 'unexpected NACK response; last successful query..' log line
wasn't really useful as it was logging the time between the last
`hwwRspNotready` and the next (and last) `hwwRspNack` responses. We actually
want to log the interval between the requests, that can cause
the Nack response (if happening with more than 500ms delay).

With this change, we use a circular buffer and keep trace of the timing
interval of the last 10 requests, to better investigate usb timing
issues.