firmware/tests: test against bitbox02 simulators
The simulators are downloaded base on the simulators.json file if the
file does not exist (or the hash is wrong).

This allows us to exercise the library much more easily without
needing to mock everything, as the simulator behaves like the real
BitBox02 for the most part.

We could automatically download all releases by using the GitHub
releases
API (https://api.github.com/repos/BitBoxSwiss/bitbox02-firmware/releases),
but the downside is that there would be network requests each time you
run the unit tests. For now we simply keep the list manually.

The simulators are only released for linux-amd64 for now, so we skip
tests if we are not running in this environment.