api/firmware/device: info use device status fields
Return struct containing REQ_INFO response data from device.info().
Amongst the already existing fields that were previously returned by
info() as a tuple, also add the initialized status to the struct that is
returned after firmware version 9.20.0. In case the device does not
respond with the initialized status byte yet it will be nil. Otherwise
true/false depending on the initialized status (seeded and backup
created) of the device.

Also use the information from device.info to set the device.status
earlier to avoid showing the password video if the device is not
initialized.