api/firmware/device: info use device status fields
Return struct containing REQ_INFO response data from device.info().
Amongst the already existing fields that were previously returned by
info() as a tuple, also add the initialized status to the struct that is
returned after firmware version 9.20.0. In case the device does not
respond with the initialized status byte yet it will be nil. Otherwise
true/false depending on the initialized status (seeded and backup
created) of the device.

Also use the information from device.info to set the device.status
earlier to avoid showing the password video if the device is not
initialized.

Also rename OP_INFO to REQ_INFO because the name changed after this commit:
https://github.com/BitBoxSwiss/bitbox02-firmware/commit/a7850125728bb447a9b79d1032bcd82c7faf83fd
The name should be the same as in the firmware repositroy so that the
code base is easier to understand.

Also rename StatusInitialized to StatusUnlocked to avoid confusion.