eth_sign_tx: integrate address case feature to go library
see : BitBoxSwiss/bitbox02-firmware@055bda6

For ETH signing legacy and EIP1559 transactions, the BitBox02 device
shows confirmation screens for recipient and amount. For recipient, the
device was previously showing the address in mixed case according to
EIP-55, no matter the case user has entered. This created confusions
among the users.

That is why not the go library detects the case of the recipient address
useer inputted and sends this case information to BitBox02 device, which
will in turn show the recipient address in the same case that it got
from the library.

Signed-off-by: asi345 <inanata15@gmail.com>