api/firmware: prevent sleep on macOS during long running queries
When doing long interactions on the device, e.g. recovering from 24
words, and macOS goes to sleep, USB interaction would be slowed down
and the workflow would be aborted.

This commit prevents this from happening by preventing macOS from
going to sleep during such an interaction.

The alternative was to expose callbacks here and let the
BitBoxApp (and other users) implement the callbacks, but it seems like
every app that uses this lib would want this. In case not, we add a
build flag if not desired.