btc: add support for detecting self-sends to different accounts
Can only make use of it starting v9.22.0.

The .proto messages have been pulled from
https://github.com/BitBoxSwiss/bitbox02-firmware/tree/c35647384740683ab06309ff34b3395d5c026d58/messages
and the .pb.go files generated from it according to this repo's README instructions.

When sending coins to an address of the same keystore, but different
account, you can provide the script config of the other account in the
`outputScriptConfigs` and link the output to it.