common: add product strings for BitBox02Plus
Allows the BitBoxApp to detect the product. These might be placeholder
values to be renamed later.

The existing product names are prefixed with BitBox02 for
disambiguation.

Standard is renamed to Multi, as that's what the product is actually
named.

HID is renamed to Device, as the string is not only used in HID
descriptors, but also in the Bluetooth characteristic.