api/firmware: make sure a series of request doesn't get interrupted
Some API calls, like BTCSign, or BluetoothUpgrade, require a series of
request<>response pairs. If the device gets an unexpected message, it
fails.

However, this is not concurrency safe - if another goroutine sends
something in the middle of it, the API call in progress and the
interrupting call both fail.

To solve this, we introduce `apiLock` that ensures a series of queries
does not get interrupted.

Most API functions only use a single query, so we acquire this lock by
default in `.query()`.

We add a `nonAtomicQuery()` variant with a `atomicQueries()` helper
function to enable making a series of calls.