firmware: add BTCSignPSBT function
PSBT is the commonly supported format for handling Bitcoin
transactions.

We use btcd's PSBT library, which currently supports PSBTv0 (BIP-174),
and not yet PSBTv2 (BIP-370).

We add a `BTCSignPSBT()` function so third party wallets can more
easily integrate this library, and so the BitBoxApp can start making
use of PSBTs internally as well. Benefits include:

- well specified, so the concepts of which metadata goes where and is
  needed for what should be easier to understand and maintain

- consistency between client libs (bitbox02-api-go, bitbox-api-rs)

- can make use of btcd's PSBT library, which contains helper functions
  to update/sign/finalize a PSBT, so we can eliminate our own code to
  generate sigScripts/witnesses

Sign options are provided for custom modifications that are not
encoded in the PSBT. We could also use PSBT proprietary fields for
them, but that seems more annoying to deal with.