psbt: infer send-to-self data from PSBT
The PSBT contains all the info to be able to tell the BitBox if an
output belongs to the same keystore, and if so, if it belongs to the
same account or a different account. Depending on which it is, we add
the scriptConfig to the input script configs or to the output script
configs.

There is some complication with maintaining backwards compatbility, as
the BitBox did not always accept non-change internal outputs or
internal outputs belonging to a different account. In these cases, the
output is confirmed as a regular outgoing payment.

findOurKey is changed to return nil if the key is not found, as that
is a different case to a parsing error in outputs.