firmware/psbt: add `BTCSignNeedsNonWitnessUTXOs()` helper
BitBox needs access to the input's previous
transactions (NonWitnessUTXO in PSBT-jargon) if not all inputs are
Taproot. Since it can be costly to retrieve them, client's can check
if they are needed using this function before providing them.

When signing using `BTCSign()` directly, one could do this with
`BTCSignNeedsPrevTxs()`, but that is not compatible with PSBT signing
workflow.

A small complication: BitBox has a bug (to be fixed in a future FW
version) where the previous transactions are required, even if all
inputs are Taproot - when an output is marked as belonging to the same
account (change or otherwise) but the output is a non-Taproot
output. The function as it is now aligns with this bug, and will be
extended/corrected once the FW with the bugfix is available. When that
happens, the `TestSimulatorBTCPSBTTaprootSpendsToNonTaproot` can be
adjusted.